+++ kuhako/src/main/java/com/capstone/kuhako/services/ClientModuleServices/PaymentMethodSelectionServiceImpl.java
@@ -1,7 +1,9 @@
 package com.capstone.kuhako.services.ClientModuleServices;
 
+import com.capstone.kuhako.models.Client;
 import com.capstone.kuhako.models.ClientModules.PaymentMethodSelection;
 import com.capstone.kuhako.repositories.ClientModuleRepository.PaymentMethodSelectionRepository;
+import com.capstone.kuhako.repositories.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,37 +14,49 @@ public class PaymentMethodSelectionServiceImpl implements PaymentMethodSelection
     @Autowired
     private PaymentMethodSelectionRepository paymentMethodSelectionRepository;
 
+    @Autowired
+    private ClientRepository clientRepository;
 
-    public void createPaymentMethodSelection(PaymentMethodSelection paymentMethodSelection){
-//      User author = userRepository.findByUsername(jwtToken.getUsernameFromToken(stringToken));
-        PaymentMethodSelection newPaymentMethodSelection = new PaymentMethodSelection();
-        newPaymentMethodSelection.setClient(paymentMethodSelection.getClient());
-        newPaymentMethodSelection.setPurchaseMethod(paymentMethodSelection.getPurchaseMethod());
-        newPaymentMethodSelection.setItemName(paymentMethodSelection.getItemName());
-        newPaymentMethodSelection.setItemPrice(paymentMethodSelection.getItemPrice());
-        newPaymentMethodSelection.setItemSpecs(paymentMethodSelection.getItemSpecs());
-        newPaymentMethodSelection.setInstallmentAmount(paymentMethodSelection.getInstallmentAmount());
-        paymentMethodSelectionRepository.save(newPaymentMethodSelection);
+    public void createPaymentMethodSelection(PaymentMethodSelection paymentMethodSelection, Long clientId){
+        Client client = clientRepository.findById(clientId).orElse(null);
+        if (client != null){
+            paymentMethodSelection.setClient(client);
+            paymentMethodSelectionRepository.save(paymentMethodSelection);
+        }
     }
     public Iterable<PaymentMethodSelection> getPaymentMethodSelection(){
         return paymentMethodSelectionRepository.findAll();
     }
 
-    public ResponseEntity deletePaymentMethodSelection(Long id){
-        paymentMethodSelectionRepository.deleteById(id);
-        return new ResponseEntity<>("PaymentMethodSelection Deleted successfully", HttpStatus.OK);
-
+    public Iterable<PaymentMethodSelection>getPaymentMethodSelectionByClientId(Long clientId){
+        return paymentMethodSelectionRepository.findPaymentMethodSelectionByClientId(clientId);
     }
 
-    public ResponseEntity updatePaymentMethodSelection(Long id, PaymentMethodSelection paymentMethodSelection){
-        PaymentMethodSelection paymentMethodSelectionForUpdate = paymentMethodSelectionRepository.findById(id).get();
-        paymentMethodSelectionForUpdate.setClient(paymentMethodSelection.getClient());
-        paymentMethodSelectionForUpdate.setPurchaseMethod(paymentMethodSelection.getPurchaseMethod());
-        paymentMethodSelectionForUpdate.setItemName(paymentMethodSelection.getItemName());
-        paymentMethodSelectionForUpdate.setItemPrice(paymentMethodSelection.getItemPrice());
-        paymentMethodSelectionForUpdate.setItemSpecs(paymentMethodSelection.getItemSpecs());
-        paymentMethodSelectionForUpdate.setInstallmentAmount(paymentMethodSelection.getInstallmentAmount());
-        paymentMethodSelectionRepository.save(paymentMethodSelectionForUpdate);
-        return new ResponseEntity("PaymentMethodSelection updated successfully", HttpStatus.OK);
+    public ResponseEntity deletePaymentMethodSelection(Long clientId,Long id){
+        PaymentMethodSelection paymentMethodSelectionToDelete = paymentMethodSelectionRepository.findById(clientId).orElse(null);
+        if(paymentMethodSelectionToDelete != null && paymentMethodSelectionToDelete.getClient().getClient_id().equals(clientId)){
+            paymentMethodSelectionRepository.deleteById(id);
+            return new ResponseEntity<>("Payment Method Deleted Successfully!",HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Payment Method not found or does not belong to the specified client", HttpStatus.NOT_FOUND);
+        }
+    }
+
+    public ResponseEntity updatePaymentMethodSelection(Long clientId, Long id, PaymentMethodSelection paymentMethodSelection){
+        PaymentMethodSelection paymentMethodSelectionForUpdate = paymentMethodSelectionRepository.findById(id).orElse(null);
+        if(paymentMethodSelectionForUpdate != null){
+            paymentMethodSelectionForUpdate.setPurchaseMethod(paymentMethodSelection.getPurchaseMethod());
+            paymentMethodSelectionForUpdate.setItemName(paymentMethodSelection.getItemName());
+            paymentMethodSelectionForUpdate.setItemPrice(paymentMethodSelection.getItemPrice());
+            paymentMethodSelectionForUpdate.setItemSpecs(paymentMethodSelection.getItemSpecs());
+            paymentMethodSelectionForUpdate.setInstallmentAmount(paymentMethodSelection.getInstallmentAmount());
+            Client client = clientRepository.findById(clientId).orElse(null);
+            if(client != null){
+                paymentMethodSelectionForUpdate.setClient(client);
+            }
+            paymentMethodSelectionRepository.save(paymentMethodSelectionForUpdate);
+            return new ResponseEntity<>("Payment Method Updated!",HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Payment Method Not Found",HttpStatus.NOT_FOUND);
     }
 }
